--- HDesign/lib/index.ts
@@ -1,3 +1,0 @@
-export {default as Button} from './components/Button/Button';
-export {default as Text} from './components/Text/Text';
-export {HDesignProvider} from './theme/HDesignProvider';
