+++ HDesign/src/lib/components/BottomSheet/BottomSheet.style.ts
