+++ HDesign/src/lib/components/FixedButton/FixedButton.type.ts
