+++ HDesign/src/lib/components/Text/Text.type.ts
