+++ HDesign/src/lib/svg.d.ts
@@ -1,0 +1,4 @@
+declare module '*.svg' {
+  const value: any;
+  export default value;
+}
